+++ example/maze/index.dart
@@ -5,7 +5,6 @@ import 'package:boarding/grid.dart';
 import 'package:boarding/util.dart';
 import 'package:boarding/boarding.dart';
 
-part 'model/grid.dart';
 part 'view/board.dart';
 
 main() {